# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "User creds to authenticate",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.Credentials"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User authenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "User creds to register",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.FullCredentials"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "task"
                ],
                "summary": "get all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "projects": {
                                "$ref": "#/definitions/task.Task"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "create task",
                "parameters": [
                    {
                        "description": "create task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.TaskDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Task created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "update task",
                "parameters": [
                    {
                        "description": "Updating task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task updated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "get by id",
                "parameters": [
                    {
                        "description": "Task id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "delete by id",
                "parameters": [
                    {
                        "description": "Task id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/projects": {
            "get": {
                "tags": [
                    "project"
                ],
                "summary": "get all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "projects": {
                                "$ref": "#/definitions/auth.Project"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "create project",
                "parameters": [
                    {
                        "description": "create request",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ProjectDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Project created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "update project",
                "parameters": [
                    {
                        "description": "Updating user",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.Project"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Project updated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "get by id",
                "parameters": [
                    {
                        "description": "Project id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.Project"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "delete by id",
                "parameters": [
                    {
                        "description": "Project id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Project deleted"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.Credentials": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "auth.FullCredentials": {
            "type": "object",
            "required": [
                "login",
                "name",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.Project": {
            "type": "object",
            "required": [
                "cost",
                "dimension",
                "id",
                "name",
                "type"
            ],
            "properties": {
                "cost": {
                    "type": "number"
                },
                "dimension": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "auth.ProjectDto": {
            "type": "object",
            "required": [
                "cost",
                "dimension",
                "name",
                "type"
            ],
            "properties": {
                "cost": {
                    "type": "number"
                },
                "dimension": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "task.Task": {
            "type": "object",
            "required": [
                "amount",
                "client_name",
                "description",
                "id",
                "project_id",
                "user_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "client_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "project_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "task.TaskDto": {
            "type": "object",
            "required": [
                "amount",
                "client_name",
                "description",
                "project_id",
                "user_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "client_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "project_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}